refactor(backends): extract xorq_datafusion and databricks impls into backend.py
Phase 2 of #2066. Both __init__.py files held full backend
implementations (980 and 670 lines) — SQL compilation, UDF registration,
type converters, connection management — instead of re-exporting from
submodules.

Move the implementations into a backend.py submodule per package and
reduce each __init__.py to a thin re-export of Backend and connect (the
only names external code imports). test_register.py imported the private
_select_and_cast helper; point it at the backend submodule.
