+++ python/xorq/backends/databricks/__init__.py
@@ -2,675 +2,10 @@
 
 from __future__ import annotations
 
-import contextlib
-import functools
-import getpass
-import json
-import os
-import sys
-import tempfile
-from pathlib import Path
-from typing import TYPE_CHECKING, Any
-
-import databricks.sql
-import pyarrow as pa
-import pyarrow_hotfix  # noqa: F401
-import sqlglot as sg
-import sqlglot.expressions as sge
-
-import xorq.common.exceptions as exc
-import xorq.vendor.ibis as ibis
-import xorq.vendor.ibis.backends.sql.compilers as sc
-import xorq.vendor.ibis.expr.datatypes as dt
-import xorq.vendor.ibis.expr.operations as ops
-import xorq.vendor.ibis.expr.schema as sch
-import xorq.vendor.ibis.expr.types as ir
-from xorq.config import default_backend
-from xorq.vendor.ibis import util
-from xorq.vendor.ibis.backends import CanCreateDatabase, UrlFromPath
-from xorq.vendor.ibis.backends.sql import SQLBackend
-from xorq.vendor.ibis.backends.sql.compilers.base import STAR, AlterTable, RenameTable
-from xorq.vendor.ibis.backends.sql.datatypes import DatabricksType
-
-
-if TYPE_CHECKING:
-    from collections.abc import Callable, Iterable, Mapping
-
-    import pandas as pd
-    import polars as pl
-
-    from xorq.vendor.ibis.expr.schema import IntoSchema
+from xorq.backends.databricks.backend import Backend, connect
 
 
 __all__ = [
     "Backend",
     "connect",
 ]
-
-
-def _databricks_type_to_ibis(typ, nullable: bool = True) -> dt.DataType:
-    """Convert a Databricks type to an Ibis type."""
-    typname = typ["name"]
-    match typname:
-        case "array":
-            return dt.Array(
-                _databricks_type_to_ibis(
-                    typ["element_type"], nullable=typ["element_nullable"]
-                ),
-                nullable=nullable,
-            )
-        case "map":
-            return dt.Map(
-                key_type=_databricks_type_to_ibis(typ["key_type"]),
-                value_type=_databricks_type_to_ibis(
-                    typ["value_type"], nullable=typ["value_nullable"]
-                ),
-                nullable=nullable,
-            )
-        case "struct":
-            return dt.Struct(
-                {
-                    field["name"]: _databricks_type_to_ibis(
-                        field["type"], nullable=field["nullable"]
-                    )
-                    for field in typ["fields"]
-                },
-                nullable=nullable,
-            )
-        case "decimal":
-            return dt.Decimal(
-                precision=typ["precision"], scale=typ["scale"], nullable=nullable
-            )
-        case _:
-            return DatabricksType.from_string(typname, nullable=nullable)
-
-
-def _databricks_schema_to_ibis(schema: Iterable[Mapping[str, Any]]) -> sch.Schema:
-    """Convert a Databricks schema to an Ibis schema."""
-    return sch.Schema(
-        {
-            item["name"]: _databricks_type_to_ibis(
-                item["type"], nullable=item["nullable"]
-            )
-            for item in schema
-        }
-    )
-
-
-def _remove_memtable_file(path: str, con) -> None:
-    """Finalizer for in-memory tables.
-
-    The view that references the storage is temporary and will be
-    automatically removed, so remove only the backing file.
-    """
-    with con.cursor() as cur:
-        cur.execute(f"REMOVE '{path}'")
-
-
-def _batch_producer(con, sql, chunk_size: int):
-    with con.cursor() as cur:
-        batched_cur = cur.execute(sql)
-        while batch := batched_cur.fetchmany_arrow(size=chunk_size):
-            yield from batch.to_batches()
-
-
-def _casting_producer(producer, schema):
-    for batch in producer:
-        yield batch.cast(schema)
-
-
-class Backend(SQLBackend, CanCreateDatabase, UrlFromPath):
-    name = "databricks"
-    compiler = sc.databricks.compiler
-
-    @property
-    def current_catalog(self) -> str:
-        with self._safe_raw_sql(sg.select(self.compiler.f.current_catalog())) as cur:
-            [(db,)] = cur.fetchall()
-        return db
-
-    @property
-    def current_database(self) -> str:
-        with self._safe_raw_sql(sg.select(self.compiler.f.current_database())) as cur:
-            [(db,)] = cur.fetchall()
-        return db
-
-    def raw_sql(self, query: str | sg.Expression, **kwargs: Any) -> Any:
-        with contextlib.suppress(AttributeError):
-            query = query.sql(self.dialect)
-        cur = self.con.cursor()
-        try:
-            cur.execute(query, **kwargs)
-        except Exception:
-            cur.close()
-            raise
-        return cur
-
-    def create_table(
-        self,
-        name: str,
-        /,
-        obj: ir.Table
-        | pd.DataFrame
-        | pa.Table
-        | pl.DataFrame
-        | pl.LazyFrame
-        | None = None,
-        *,
-        schema: IntoSchema | None = None,
-        database: str | None = None,
-        temp: bool = False,
-        overwrite: bool = False,
-        using: str = "delta",
-        location: str | None = None,
-        tblproperties: Mapping[str, str] | None = None,
-    ):
-        """Create a table in Databricks.
-
-        Parameters
-        ----------
-        name
-            Name of the table to create
-        obj
-            The data with which to populate the table; optional, but at least
-            one of `obj` or `schema` must be specified
-        schema
-            The schema of the table to create; optional, but at least one of
-            `obj` or `schema` must be specified
-        database
-            The name of the database in which to create the table; if not
-            passed, the current database is used.
-
-            For multi-level table hierarchies, you can pass in a dotted string
-            path like `"catalog.database"` or a tuple of strings like
-            `("catalog", "database")`.
-        temp
-            Create a temporary table
-        overwrite
-            If `True`, replace the table if it already exists, otherwise fail
-            if the table exists
-        using
-            Data source format
-        location
-            Storage location for the table
-        tblproperties
-            Table properties
-        """
-        if temp:
-            raise exc.UnsupportedOperationError("Temporary tables not yet supported")
-
-        table_loc = self._to_sqlglot_table(database)
-
-        catalog = table_loc.catalog or self.current_catalog
-        database = table_loc.db or self.current_database
-
-        if obj is None and schema is None:
-            raise ValueError("Either `obj` or `schema` must be specified")
-        if schema is not None:
-            schema = ibis.schema(schema)
-
-        properties = [sge.FileFormatProperty(this=self.compiler.v[using.upper()])]
-
-        if location is not None:
-            properties.append(sge.LocationProperty(this=sge.convert(location)))
-
-        for key, value in (tblproperties or {}).items():
-            properties.append(
-                sge.Property(this=sge.convert(str(key)), value=sge.convert(str(value)))
-            )
-
-        if obj is not None:
-            if not isinstance(obj, ir.Expr):
-                table = ibis.memtable(obj)
-            else:
-                table = obj
-
-            if not schema:
-                schema = table.schema()
-
-            self._run_pre_execute_hooks(table)
-
-            query = self.compiler.to_sqlglot(table)
-        else:
-            query = None
-
-        if overwrite:
-            temp_name = util.gen_name("databricks_table")
-        else:
-            temp_name = name
-
-        quoted = self.compiler.quoted
-        dialect = self.dialect
-
-        initial_table = sg.table(temp_name, catalog=catalog, db=database, quoted=quoted)
-        target = sge.Schema(
-            this=initial_table,
-            expressions=schema.to_sqlglot(dialect),
-        )
-
-        properties = sge.Properties(expressions=properties)
-        create_stmt = sge.Create(kind="TABLE", this=target, properties=properties)
-
-        # This is the same table as initial_table unless overwrite == True
-        final_table = sg.table(name, catalog=catalog, db=database, quoted=quoted)
-        with self._safe_raw_sql(create_stmt) as cur:
-            if query is not None:
-                insert_stmt = sge.insert(query, into=initial_table).sql(dialect)
-                cur.execute(insert_stmt).fetchall()
-
-            if overwrite:
-                cur.execute(
-                    sge.Drop(kind="TABLE", this=final_table, exists=True).sql(dialect)
-                )
-                if temp:
-                    cur.execute(
-                        sge.Create(
-                            kind="TABLE",
-                            this=final_table,
-                            expression=sg.select(STAR).from_(initial_table),
-                            properties=properties,
-                        ).sql(dialect)
-                    )
-                    cur.execute(
-                        sge.Drop(kind="TABLE", this=initial_table, exists=True).sql(
-                            dialect
-                        )
-                    )
-                else:
-                    cur.execute(
-                        AlterTable(
-                            this=initial_table, actions=[RenameTable(this=final_table)]
-                        ).sql(dialect)
-                    )
-
-        return self.table(name, database=(catalog, database))
-
-    def table(self, name: str, /, *, database: str | None = None) -> ir.Table:
-        table_loc = self._to_sqlglot_table(database)
-
-        catalog = table_loc.catalog or None
-        database = table_loc.db or None
-
-        table_schema = self.get_schema(name, catalog=catalog, database=database)
-        return ops.DatabaseTable(
-            name,
-            schema=table_schema,
-            source=self,
-            namespace=ops.Namespace(catalog=catalog, database=database),
-        ).to_expr()
-
-    def get_schema(
-        self,
-        table_name: str,
-        *,
-        catalog: str | None = None,
-        database: str | None = None,
-    ) -> sch.Schema:
-        """Compute the schema of a `table`.
-
-        Parameters
-        ----------
-        table_name
-            May **not** be fully qualified. Use `database` if you want to
-            qualify the identifier.
-        catalog
-            Catalog name
-        database
-            Database name
-
-        Returns
-        -------
-        sch.Schema
-            Ibis schema
-        """
-        table = sg.table(
-            table_name, db=database, catalog=catalog, quoted=self.compiler.quoted
-        )
-        view_name = sg.to_identifier(
-            util.gen_name(f"databricks_view_of_{table_name}"),
-            quoted=self.compiler.quoted,
-        )
-        view = sge.Create(
-            kind="VIEW",
-            this=view_name,
-            expression=sg.select(STAR).from_(table),
-            properties=sge.Properties(expressions=[sge.TemporaryProperty()]),
-        )
-        view_sql = view.sql(self.dialect)
-        try:
-            with self.con.cursor() as cur:
-                # create a temporary view to get the schema
-                #
-                # this is necessary to support streaming tables, which don't
-                # directly support using
-                #
-                # DESCRIBE EXTENDED {streaming_table} AS JSON
-                #
-                # the create-a-view-then-drop-it strategy works for all table types
-                cur.execute(view_sql)
-        except databricks.sql.exc.ServerOperationError as e:
-            raise exc.TableNotFound(
-                f"Table {table_name!r} not found in "
-                f"{catalog or self.current_catalog}.{database or self.current_database}"
-            ) from e
-        try:
-            with self.con.cursor() as cur:
-                [(out,)] = cur.execute(
-                    f"DESCRIBE EXTENDED {view_name.sql(self.dialect)} AS JSON"
-                ).fetchall()
-        finally:
-            # clean up the temporary view
-            with self.con.cursor() as cur:
-                cur.execute(
-                    sge.Drop(this=view_name, kind="VIEW", exists=True).sql(self.dialect)
-                )
-
-        js = json.loads(out)
-        return _databricks_schema_to_ibis(js["columns"])
-
-    @contextlib.contextmanager
-    def _safe_raw_sql(self, query, *args, **kwargs):
-        with contextlib.suppress(AttributeError):
-            query = query.sql(self.dialect)
-        with self.con.cursor() as cur:
-            yield cur.execute(query, *args, **kwargs)
-
-    def list_catalogs(self, *, like: str | None = None) -> list[str]:
-        with self.con.cursor() as cur:
-            out = cur.catalogs().fetchall_arrow()
-        return self._filter_with_like(out["TABLE_CAT"].to_pylist(), like)
-
-    def list_databases(
-        self, *, like: str | None = None, catalog: str | None = None
-    ) -> list[str]:
-        with self.con.cursor() as cur:
-            out = cur.schemas(
-                catalog_name=catalog or self.current_catalog
-            ).fetchall_arrow()
-        return self._filter_with_like(out["TABLE_SCHEM"].to_pylist(), like=like)
-
-    @functools.cached_property
-    def version(self) -> str:
-        query = sg.select(self.compiler.f.current_version())
-        with self._safe_raw_sql(query) as cur:
-            [(version_info,)] = cur.fetchall()
-        return version_info["dbsql_version"]
-
-    def do_connect(
-        self,
-        *,
-        server_hostname: str | None = None,
-        http_path: str | None = None,
-        access_token: str | None = None,
-        auth_type: str | None = None,
-        credentials_provider: str | None = None,
-        password: str | None = None,
-        username: str | None = None,
-        session_configuration: Mapping[str, str] | None = None,
-        http_headers: list[tuple[str, str]] | None = None,
-        catalog: str | None = None,
-        schema: str = "default",
-        use_cloud_fetch: bool = False,
-        memtable_volume: str | None = None,
-        staging_allowed_local_path: str | None = None,
-        **config: Any,
-    ) -> None:
-        """Create an xorq client connected to a Databricks cloud instance."""
-        if staging_allowed_local_path is None:
-            staging_allowed_local_path = tempfile.gettempdir()
-        self._server_hostname = server_hostname or os.environ.get(
-            "DATABRICKS_SERVER_HOSTNAME"
-        )
-        self._http_path = http_path or os.environ.get("DATABRICKS_HTTP_PATH")
-        self._access_token = access_token or os.environ.get("DATABRICKS_TOKEN")
-        self.con = databricks.sql.connect(
-            server_hostname=self._server_hostname,
-            http_path=self._http_path,
-            access_token=self._access_token,
-            auth_type=auth_type,
-            credentials_provider=credentials_provider,
-            password=password,
-            username=username,
-            session_configuration=session_configuration,
-            http_headers=http_headers,
-            catalog=catalog,
-            schema=schema,
-            use_cloud_fetch=use_cloud_fetch,
-            staging_allowed_local_path=staging_allowed_local_path,
-            **config,
-        )
-        if memtable_volume is None:
-            short_version = "".join(map(str, sys.version_info[:3]))
-            memtable_volume = (
-                f"{getpass.getuser()}-py={short_version}-pid={os.getpid()}"
-            )
-        self._memtable_volume = memtable_volume
-        self._memtable_catalog = self.current_catalog
-        self._memtable_database = self.current_database
-        self._post_connect(memtable_volume=memtable_volume)
-
-    @contextlib.contextmanager
-    def begin(self):
-        with self.con.cursor() as cur:
-            yield cur
-
-    @util.experimental
-    @classmethod
-    def from_connection(cls, con, /, *, memtable_volume: str | None = None) -> Backend:
-        """Create an xorq client from an existing connection to a Databricks cloud instance.
-
-        Parameters
-        ----------
-        con
-            An existing connection to a Databricks database.
-        memtable_volume
-            The volume to use for xorq memtables.
-        """
-        new_backend = cls()
-        new_backend._can_reconnect = False
-        new_backend.con = con
-        new_backend._post_connect(memtable_volume=memtable_volume)
-        return new_backend
-
-    def _post_connect(self, *, memtable_volume: str) -> None:
-        sql = f"CREATE VOLUME IF NOT EXISTS `{memtable_volume}` COMMENT 'xorq memtable storage volume'"
-        with self.con.cursor() as cur:
-            cur.execute(sql)
-
-    @functools.cached_property
-    def _memtable_volume_path(self) -> str:
-        return f"/Volumes/{self._memtable_catalog}/{self._memtable_database}/{self._memtable_volume}"
-
-    def _register_in_memory_table(self, op: ops.InMemoryTable) -> None:
-        import pyarrow.parquet as pq  # noqa: PLC0415
-
-        quoted = self.compiler.quoted
-        name = op.name
-        stem = f"{name}.parquet"
-
-        upstream_path = f"{self._memtable_volume_path}/{stem}"
-        sql = sge.Create(
-            kind="VIEW",
-            this=sg.table(name, quoted=quoted),
-            expression=sg.select(STAR).from_(
-                sg.table(upstream_path, db="parquet", quoted=quoted)
-            ),
-            properties=sge.Properties(expressions=[sge.TemporaryProperty()]),
-        ).sql(self.dialect)
-        data = op.data.to_pyarrow(schema=op.schema)
-        with tempfile.TemporaryDirectory(ignore_cleanup_errors=True) as tmpdir:
-            path = Path(tmpdir, stem)
-            put_into = f"PUT '{path}' INTO '{upstream_path}' OVERWRITE"
-            # optimize for bandwidth: zstd typically compresses better
-            pq.write_table(data, path, compression="zstd")
-            with self.con.cursor() as cur:
-                cur.execute(put_into)
-                cur.execute(sql)
-
-    def read_record_batches(
-        self,
-        record_batches: pa.RecordBatchReader,
-        table_name: str | None = None,
-        mode: str = "create",
-        **kwargs: Any,
-    ) -> ir.Table:
-        from xorq.common.utils.databricks_utils import DatabricksADBC  # noqa: PLC0415
-
-        table_name = table_name or util.gen_name("databricks_record_batches")
-        DatabricksADBC(self).adbc_ingest(
-            table_name, record_batches, mode=mode, **kwargs
-        )
-        return self.table(table_name)
-
-    def read_parquet(
-        self, path: str | Path, table_name: str | None = None, **kwargs: Any
-    ) -> ir.Table:
-        if table_name is None:
-            table_name = util.gen_name("ls-read-parquet")
-        record_batches = default_backend().read_parquet(path).to_pyarrow_batches()
-        return self.read_record_batches(
-            record_batches=record_batches,
-            table_name=table_name,
-            **kwargs,
-        )
-
-    def _make_memtable_finalizer(self, name: str) -> Callable[..., None]:
-        path = f"{self._memtable_volume_path}/{name}.parquet"
-        return functools.partial(_remove_memtable_file, path, self.con)
-
-    def create_database(
-        self, name: str, /, *, catalog: str | None = None, force: bool = False
-    ) -> None:
-        name = sg.table(name, catalog=catalog, quoted=self.compiler.quoted)
-        with self._safe_raw_sql(sge.Create(this=name, kind="SCHEMA", replace=force)):
-            pass
-
-    def drop_database(
-        self, name: str, /, *, catalog: str | None = None, force: bool = False
-    ) -> None:
-        name = sg.table(name, catalog=catalog, quoted=self.compiler.quoted)
-        with self._safe_raw_sql(sge.Drop(this=name, kind="SCHEMA", replace=force)):
-            pass
-
-    def list_tables(
-        self, *, like: str | None = None, database: tuple[str, str] | str | None = None
-    ) -> list[str]:
-        table_loc = self._to_sqlglot_table(database)
-
-        catalog = table_loc.catalog or self.current_catalog
-        database = table_loc.db or self.current_database
-
-        with self.con.cursor() as cur:
-            cur.tables(catalog_name=catalog, schema_name=database)
-            out = cur.fetchall_arrow()
-
-        return self._filter_with_like(out["TABLE_NAME"].to_pylist(), like)
-
-    def to_pyarrow_batches(
-        self,
-        expr: ir.Expr,
-        /,
-        *,
-        params: Mapping[ir.Scalar, Any] | None = None,
-        limit: int | str | None = None,
-        chunk_size: int = 1_000_000,
-        **_: Any,
-    ) -> pa.ipc.RecordBatchReader:
-        """Return a stream of record batches.
-
-        Parameters
-        ----------
-        expr
-            Ibis expression
-        params
-            Bound parameters
-        limit
-            Limit the result to this number of rows
-        chunk_size
-            The number of rows to fetch per batch
-        """
-        self._run_pre_execute_hooks(expr)
-        table = expr.as_table()
-        sql = self.compile(table, limit=limit, params=params)
-
-        pyarrow_schema = expr.as_table().schema().to_pyarrow()
-        producer = _casting_producer(
-            _batch_producer(self.con, sql, chunk_size), pyarrow_schema
-        )
-        return pa.ipc.RecordBatchReader.from_batches(pyarrow_schema, producer)
-
-    def to_pyarrow(
-        self,
-        expr: ir.Expr,
-        /,
-        *,
-        params: Mapping[ir.Scalar, Any] | None = None,
-        limit: int | str | None = None,
-        **kwargs: Any,
-    ) -> pa.Table:
-        self._run_pre_execute_hooks(expr)
-
-        sql = self.compile(expr, limit=limit, params=params, **kwargs)
-        with self._safe_raw_sql(sql) as cur:
-            res = cur.fetchall_arrow()
-
-        target_schema = expr.as_table().schema().to_pyarrow()
-        if res is None:
-            res = target_schema.empty_table()
-
-        return expr.__pyarrow_result__(res)
-
-    def _fetch_from_cursor(self, cursor, schema: sch.Schema) -> pd.DataFrame:
-        if (table := cursor.fetchall_arrow()) is None:
-            table = schema.to_pyarrow().empty_table()
-        df = table.to_pandas(timestamp_as_object=True)
-        df.columns = list(schema.names)
-        return df
-
-    def _get_schema_using_query(self, query: str) -> sch.Schema:
-        with self._safe_raw_sql(
-            sge.Describe(this=sg.parse_one(query, read=self.dialect))
-        ) as cur:
-            rows = cur.fetchall_arrow()
-
-        rows = rows.to_pydict()
-
-        type_mapper = self.compiler.type_mapper
-        return sch.Schema(
-            {
-                name: type_mapper.from_string(typ, nullable=True)
-                for name, typ in zip(rows["col_name"], rows["data_type"])
-            }
-        )
-
-    def rename_table(self, old_name: str, new_name: str) -> None:
-        """Rename an existing table.
-
-        Parameters
-        ----------
-        old_name
-            The old name of the table.
-        new_name
-            The new name of the table.
-        """
-        old = sg.table(old_name, quoted=True)
-        new = sg.table(new_name, quoted=True)
-        query = AlterTable(
-            this=old, exists=False, actions=[RenameTable(this=new, exists=True)]
-        )
-        with self._safe_raw_sql(query):
-            pass
-
-    @classmethod
-    def connect_env(cls, **kwargs):
-        from xorq.common.utils.databricks_utils import make_connection  # noqa: PLC0415
-
-        return make_connection(**kwargs)
-
-    _top_level_methods = ("connect_env",)
-
-
-def connect(*args, **kwargs):
-    con = Backend()
-    con.reconnect(*args, **kwargs)
-    return con
